Moved pages and componnents to respective folders + Integrated Newsletter frontend
