+++ client/src/App.js
@@ -2,9 +2,8 @@ import { Route, Routes } from "react-router-dom";
 import "./App.css";
 import AboutPratibhaNakshatra from "./AboutPratibhaNakshatra";
 import Donations from "./Donations";
-import ThankYou from "./ThankYou";
-import Newsletter from "./Newsletter";
-import Video from "./Video";
+import ThankYou from "./pages/ThankYou";
+import Video from "./components/Video";
 import Home from "./pages/Home";
 import Contact from "./Contact";
 
@@ -17,7 +16,6 @@ const App = () => (
 		<Route path="/donations" element={<Donations />} />
 		<Route path="/thank-you" element={<ThankYou /> }  />
 		<Route path="/contact" element={<Contact />} />
-		<Route path="/newsletter" element={<Newsletter />} />
 	</Routes>
   <Video />
   </>
