+++ client/src/pages/Home.js
@@ -3,7 +3,7 @@
 import Header from "../components/Header.js";
 import "./home.css";
 import Footer from "../components/Footer.js";
-
+import Newsletter from "../components/Newsletter";
 
 import homepageImage from "../images/homepage-image.jpeg";
 // import bgVideo from "../videos/bg-video.mp4";
@@ -49,6 +49,7 @@ export function Home() {
 					preserve, maintain and handover to their newer generation.
 				</p>
 			</div>
+			<Newsletter />
 			<Footer />
 		</main>
 	);
